requirements.txt
wikipedia
inflect
apocrypha
pyyaml
types-PyYaml
types-beautifulsoup4
lxml

mypy
ruff
